+++ client/src/pages/automation/project-instances/components/ProjectInstanceWorkflowList.tsx
@@ -1,6 +1,6 @@
 import {ProjectInstanceWorkflowModel} from '@/middleware/helios/configuration';
 import {ComponentDefinitionBasicModel} from '@/middleware/hermes/configuration';
-import ProjectInstanceWorkflowListItem from '@/pages/automation/project-instances/ProjectInstanceWorkflowListItem';
+import ProjectInstanceWorkflowListItem from '@/pages/automation/project-instances/components/ProjectInstanceWorkflowListItem';
 import {useGetProjectWorkflowsQuery} from '@/queries/projects.queries';
 import {useGetTaskDispatcherDefinitionsQuery} from '@/queries/taskDispatcherDefinitions.queries';
 import {useGetComponentDefinitionsQuery} from 'queries/componentDefinitions.queries';
